third_party/WebKit/LayoutTests/fast/parser/noscript-with-javascript-disabled.html

<html>

<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpChildFrames();
    testRunner.overridePreference("WebKitJavaScriptEnabled", false);
}
</script>

<noscript> 
<body>
    FAIL: This content should not render because whether scripting is enabled
    (for the purposes of the 'noscript' tag) is locked in when the parser starts.
</noscript>

The text inside the 'noscript' tag inside the iframe should render:

<iframe
    src="data:text/html,<noscript><body>PASS: This test case verifies that contents inside &lt;noscript&gt; are rendered when Javascript is disabled.</body></noscript>">
</iframe>

</body>
</html>
