third_party/blink/web_tests/platform/generic/fast/parser/noscript-with-javascript-disabled-expected.txt

The text inside the 'noscript' tag inside the iframe should render: 

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS: This test case verifies that contents inside are rendered when Javascript is disabled.
